+++ Assets/Scripts/AttributeManager.cs
@@ -1,0 +1,11 @@
+using System.Collections;
+using System.Collections.Generic;
+using UnityEngine;
+
+public class AttributeManager : MonoBehaviour
+{
+    public int health;
+    public int damage;
+
+
+}
