+++ Assets/Scripts/AttributeManager.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 0e951e7d2a3db5b4fb2868049b56a22b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
