+++ Assets/Scripts/Limb.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c8e0cc2ac083b4c4fa004ec882ad4da4
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
