+++ Assets/Weapons/Modern Guns - Handgun/_Demo Assets/SimpleShoot.cs
@@ -18,7 +18,7 @@ public class SimpleShoot : MonoBehaviour
 
     [Header("Settings")]
     [Tooltip("Specify time to destory the casing object")] [SerializeField] private float destroyTimer = 2f;
-    [Tooltip("Bullet Speed")] [SerializeField] private float shotPower = 500f;
+    [Tooltip("Bullet Speed")] [SerializeField] private float shotPower = 1000f;
     [Tooltip("Casing Ejection Speed")] [SerializeField] private float ejectPower = 150f;
 
     public AudioSource source;
@@ -94,8 +94,19 @@ public class SimpleShoot : MonoBehaviour
         if (!bulletPrefab)
         { return; }
 
+        bulletPrefab.tag = "Weapon";
+
         // Create a bullet and add force on it in direction of the barrel
         Instantiate(bulletPrefab, barrelLocation.position, barrelLocation.rotation).GetComponent<Rigidbody>().AddForce(barrelLocation.forward * shotPower);
+
+        //Code to trigger the start of the dismemberment
+        RaycastHit hit;
+
+        if (Physics.Raycast(barrelLocation.position, barrelLocation.forward, out hit, Mathf.Infinity))
+        {
+            Limb limb = hit.transform.GetComponent<Limb>();
+            limb.GetHit();
+        }
 
     }
