+++ Assets/Weapons/Modern Guns - Handgun/_Prefabs/45ACP Bullet_Head.prefab
@@ -101,6 +101,11 @@ PrefabInstance:
       propertyPath: m_TagString
       value: 'Weapon '
       objectReference: {fileID: 0}
+    - target: {fileID: 8403992145370945202, guid: 9e0ef2cf9cfc11c46bbe7f28b272f6fb,
+        type: 3}
+      propertyPath: m_CollisionDetection
+      value: 2
+      objectReference: {fileID: 0}
     - target: {fileID: 8535412081993147290, guid: 9e0ef2cf9cfc11c46bbe7f28b272f6fb,
         type: 3}
       propertyPath: m_RootOrder
