--- Assets/Zombie.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 62417e3371a63fd4ba62e262b6c83312
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
